fix: flashing permission screen
The permission screen would flash before the file
chooser appeared due to a race condition for collecting
the state. Now we have an "uninitialized" state that
we do not collect on as the default value. The initial
state is still decided by the viewModel, but we won't
collect the default value passed to the private state's
constructor.