fix: no files found screen flashing
The screen that informs users that they have no drawing files
was always appearing for a split second when loading the file
browser. Now we have a "nullDirectory" constant that can be
checked against for when there are actually no files vs when
the directory just has not yet been returned.